Move plotting to separate files to allow backend plotly use
