+++ covasim/plotting.py
@@ -1,0 +1,229 @@
+import covasim as cv
+import numpy as np
+import sciris as sc
+import plotly.graph_objects as go
+
+def get_individual_states(sim):
+    people = sim.people
+
+    states = [
+        {'name': 'Healthy',
+         'quantity': None,
+         'color': '#b9d58a',
+         'value': 0
+         },
+        {'name': 'Exposed',
+         'quantity': 'date_exposed',
+         'color': '#e37c30',
+         'value': 2
+         },
+        {'name': 'Infectious',
+         'quantity': 'date_infectious',
+         'color': '#c35d86',
+         'value': 3
+         },
+        {'name': 'Recovered',
+         'quantity': 'date_recovered',
+         'color': '#799956',
+         'value': 4
+         },
+        {'name': 'Dead',
+         'quantity': 'date_dead',
+         'color': '#000000',
+         'value': 5
+         },
+    ]
+
+    z = np.zeros((len(people), sim.npts))
+    for state in states:
+        date = state['quantity']
+        if date is not None:
+            inds = cv.defined(people[date])
+            for ind in inds:
+                z[ind, int(people[date][ind]):] = state['value']
+
+    return z, states
+
+def standard_plots(sim):
+    ''' Main simulation results '''
+    plots = []
+    to_plot = sc.dcp(cv.default_sim_plots)
+    for p,title,keylabels in to_plot.enumitems():
+        fig = go.Figure()
+        for key in keylabels:
+            label = sim.results[key].name
+            this_color = sim.results[key].color
+            y = sim.results[key][:]
+            fig.add_trace(go.Scatter(x=sim.results['t'][:], y=y, mode='lines', name=label, line_color=this_color))
+            if sim.data is not None and key in sim.data:
+                data_t = (sim.data.index-sim['start_day'])/np.timedelta64(1,'D')
+                print(sim.data.index, sim['start_day'], np.timedelta64(1,'D'), data_t)
+                ydata = sim.data[key]
+                fig.add_trace(go.Scatter(x=data_t, y=ydata, mode='markers', name=label + ' (data)', line_color=this_color))
+
+        if sim['interventions']:
+            for interv in sim['interventions']:
+                if hasattr(interv, 'days'):
+                    for interv_day in interv.days:
+                        if interv_day > 0 and interv_day < sim['n_days']:
+                            fig.add_shape(dict(type="line", xref="x", yref="paper", x0=interv_day, x1=interv_day, y0=0, y1=1, name='Intervention', line=dict(width=0.5, dash='dash')))
+                            fig.update_layout(annotations=[dict(x=interv_day, y=1.07, xref="x", yref="paper", text="Intervention change", showarrow=False)])
+
+        fig.update_layout(title={'text':title}, xaxis_title='Day', yaxis_title='Count', autosize=True)
+
+        plots.append(fig)
+    return plots
+
+def plot_people(sim) -> dict:
+    z, states = get_individual_states(sim)
+
+    fig = go.Figure()
+
+    for state in states[::-1]:  # Reverse order for plotting
+        fig.add_trace(go.Scatter(
+            x=sim.tvec, y=(z == state['value']).sum(axis=0),
+            stackgroup='one',
+            line=dict(width=0.5, color=state['color']),
+            fillcolor=state['color'],
+            hoverinfo="y+name",
+            name=state['name']
+        ))
+
+    if sim['interventions']:
+        for interv in sim['interventions']:
+                if hasattr(interv, 'days'):
+                    for interv_day in interv.days:
+                        if interv_day > 0 and interv_day < sim['n_days']:
+                            fig.add_shape(dict(type="line", xref="x", yref="paper", x0=interv_day, x1=interv_day, y0=0, y1=1, name='Intervention', line=dict(width=0.5, dash='dash')))
+                            fig.update_layout(annotations=[dict(x=interv_day, y=1.07, xref="x", yref="paper", text="Intervention change", showarrow=False)])
+
+    fig.update_layout(yaxis_range=(0, sim.n))
+    fig.update_layout(title={'text': 'Numbers of people by health state'}, xaxis_title='Day', yaxis_title='People', autosize=True)
+
+    return fig
+
+
+def animate_people(sim) -> dict:
+    z, states = get_individual_states(sim)
+
+    min_color = min(states, key=lambda x: x['value'])['value']
+    max_color = max(states, key=lambda x: x['value'])['value']
+    colorscale = [[x['value'] / max_color, x['color']] for x in states]
+
+    aspect = 3
+    y_size = int(np.ceil((z.shape[0] / aspect) ** 0.5))
+    x_size = int(np.ceil(aspect * y_size))
+
+    z = np.pad(z, ((0, x_size * y_size - z.shape[0]), (0, 0)), mode='constant', constant_values=np.nan)
+
+    days = sim.tvec
+
+    fig_dict = {
+        "data": [],
+        "layout": {},
+        "frames": []
+    }
+
+    fig_dict["layout"]["updatemenus"] = [
+        {
+            "buttons": [
+                {
+                    "args": [None, {"frame": {"duration": 200, "redraw": True},
+                                    "fromcurrent": True}],
+                    "label": "Play",
+                    "method": "animate"
+                },
+                {
+                    "args": [[None], {"frame": {"duration": 0, "redraw": True},
+                                      "mode": "immediate",
+                                      "transition": {"duration": 0}}],
+                    "label": "Pause",
+                    "method": "animate"
+                }
+            ],
+            "direction": "left",
+            "pad": {"r": 10, "t": 87},
+            "showactive": False,
+            "type": "buttons",
+            "x": 0.1,
+            "xanchor": "right",
+            "y": 0,
+            "yanchor": "top"
+        }
+    ]
+
+    sliders_dict = {
+        "active": 0,
+        "yanchor": "top",
+        "xanchor": "left",
+        "currentvalue": {
+            "font": {"size": 20},
+            "prefix": "Day: ",
+            "visible": True,
+            "xanchor": "right"
+        },
+        "transition": {"duration": 200},
+        "pad": {"b": 10, "t": 50},
+        "len": 0.9,
+        "x": 0.1,
+        "y": 0,
+        "steps": []
+    }
+
+    # make data
+    fig_dict["data"] = [go.Heatmap(z=np.reshape(z[:, 0], (y_size, x_size)),
+                                   zmin=min_color,
+                                   zmax=max_color,
+                                   colorscale=colorscale,
+                                   showscale=False,
+                                   )]
+
+    for state in states:
+        fig_dict["data"].append(go.Scatter(x=[None], y=[None], mode='markers',
+                                           marker=dict(size=10, color=state['color']),
+                                           showlegend=True, name=state['name']))
+
+    # make frames
+    for i, day in enumerate(days):
+        frame = {"data": [go.Heatmap(z=np.reshape(z[:, i], (y_size, x_size)))],
+                 "name": i}
+        fig_dict["frames"].append(frame)
+        slider_step = {"args": [
+            [i],
+            {"frame": {"duration": 5, "redraw": True},
+             "mode": "immediate", }
+        ],
+            "label": i,
+            "method": "animate"}
+        sliders_dict["steps"].append(slider_step)
+
+    fig_dict["layout"]["sliders"] = [sliders_dict]
+
+    fig = go.Figure(fig_dict)
+
+    fig.update_layout(
+    autosize=True,
+        xaxis=dict(
+            automargin=True,
+            range=[-0.5, x_size + 0.5],
+            constrain="domain",
+            showgrid=False,
+            showline=False,
+            showticklabels=False,
+        ),
+        yaxis=dict(
+            automargin=True,
+            range=[-0.5, y_size + 0.5],
+            constrain="domain",
+            scaleanchor="x",
+            scaleratio=1,
+            showgrid=False,
+            showline=False,
+            showticklabels=False,
+        ),
+    )
+
+
+    fig.update_layout(title={'text': 'Epidemic over time'})
+
+    return fig
