+++ examples/run_sim.py
@@ -44,3 +44,6 @@ sim.run(verbose=verbose)
 if do_plot:
     print('Plotting...')
     fig = sim.plot(do_save=do_save, do_show=do_show, fig_path=fig_path)
+
+fig = cv.plot_people(sim)
+fig.show(renderer='browser')
